Add callback function to GraphSearch
The callback function in GraphSearch is called whenever an Event happens.
This can bei relaxation, or current node is being processed, etc...
The callback function in GraphSearch hands the event and the node to a lambda
function which processes it. This lambda function can be provided from outside
to change functionality of the callback.