Lock search thread with L unlock with U and set step lock with P keys
When L is pressed, the thread running the search algorithm is blocked in the
callback function. By pressing U the thread can resume. When The P key is
pressed, the GraphSearch runs in lock step modus, the thread blocks every time
when callback is called and can be watched step by step. By pressing P again,
the algorithm runs normally again and can be blocked/unblocked again with L
and U keys.