Run AStar correctly by using correct units!
Unit of vertex position and path length was not the same. For the path length,
the unit was in squares, so left/right/up/down step had cost 1 and diagonal
step had cost ~1.4. But the heuristic used pixel coordinates which was much
higher (dx, dy of 16 pixels) so the heuristic dominated the cost function and
A* ran like greedy Best-First-Search. Fixed this by calculating path cost
with pixel units.